chore: improve API and add missing docs
