+++ Demo/Common/IntegrationExampleDropIn.swift
@@ -61,23 +61,6 @@ extension IntegrationExample {
 
     // MARK: - Payment response handling
 
-    internal func paymentResponseHandler(result: Result<PaymentsResponse, Error>) {
-        switch result {
-        case let .success(response):
-            if let action = response.action {
-                handle(action)
-            } else if let order = response.order,
-                      let remainingAmount = order.remainingAmount,
-                      remainingAmount.value > 0 {
-                handle(order)
-            } else {
-                finish(with: response)
-            }
-        case let .failure(error):
-            finish(with: error)
-        }
-    }
-
     internal func handle(_ order: PartialPaymentOrder) {
         requestPaymentMethods(order: order) { [weak self] paymentMethods in
             self?.handle(order, paymentMethods)
